// src/app/models/holding-info.model.ts
export class HoldingInfo implements IHoldingInfo {
  '52WeekChange': number;
  SandP52WeekChange: number;
  address1: string;
  address2: string;
  algorithm: string;
  annualHoldingsTurnover: string;
  annualReportExpenseRatio: string;
  ask: number;
  askSize: number;
  averageDailyVolume10Day: number;
  averageVolume: number;
  averageVolume10days: number;
  beta: number;
  beta3Year: number;
  bid: number;
  bidSize: number;
  bookValue: number;
  category: string;
  circulatingSupply: number;
  city: string;
  companyOfficers: Array<string>;
  country: string;
  currency: string;
  dateShortInterest: number;
  dayHigh: number;
  dayLow: number;
  dividendRate: number;
  dividendYield: number;
  earningsQuarterlyGrowth: number;
  enterpriseToEbitda: number;
  enterpriseToRevenue: number;
  enterpriseValue: number;
  exDividendDate: number;
  exchange: string;
  exchangeTimezoneName: string;
  exchangeTimezoneShortName: string;
  expireDate: string;
  fiftyDayAverage: number;
  fiftyTwoWeekHigh: number;
  fiftyTwoWeekLow: number;
  fiveYearAverageReturn: number;
  fiveYearAvgDividendYield: number;
  floatShares: number;
  forwardEps: number;
  forwardPE: number;
  fromCurrency: string;
  fullTimeEmployees: number;
  fundFamily: string;
  fundInceptionDate: string;
  gmtOffSetMilliseconds: number;
  heldPercentInsiders: number;
  heldPercentInstitutions: number;
  industry: string;
  isEsgPopulated: boolean;
  lastCapGain: number;
  lastDividendValue: number;
  lastFiscalYearEnd: number;
  lastMarket: string;
  lastSplitDate: number;
  lastSplitFactor: string;
  legalType: string;
  'logo_url': string;
  longBusinessSummary: string;
  longName: string;
  market: string;
  marketCap: number;
  maxAge: number;
  maxSupply: number;
  messageBoardId: string;
  morningStarOverallRating: number;
  morningStarRiskRating: number;
  mostRecentQuarter: number;
  navPrice: number;
  netIncomeToCommon: number;
  nextFiscalYearEnd: number;
  open: number;
  openInterest: number;
  payoutRatio: number;
  pegRatio: number;
  phone: string;
  previousClose: number;
  priceHint: number;
  priceToBook: number;
  priceToSalesTrailing12Months: number;
  profitMargins: number;
  quoteType: string;
  regularMarketDayHigh: number;
  regularMarketDayLow: number;
  regularMarketOpen: number;
  regularMarketPreviousClose: number;
  regularMarketPrice: number;
  regularMarketVolume: number;
  revenueQuarterlyGrowth: number;
  sector: string;
  sharesOutstanding: number;
  sharesPercentSharesOut: number;
  sharesShort: number;
  sharesShortPreviousMonthDate: number;
  sharesShortPriorMonth: number;
  shortName: string;
  shortPercentOfFloat: number;
  shortRatio: number;
  startDate: string;
  state: string;
  strikePrice: number;
  symbol: string;
  threeYearAverageReturn: number;
  toCurrency: string;
  totalAssets: number;
  tradeable: boolean;
  trailingAnnualDividendRate: number;
  trailingAnnualDividendYield: number;
  trailingEps: number;
  trailingPE: number;
  twoHundredDayAverage: number;
  volume: number;
  volume24Hr: number;
  volumeAllCurrencies: number;
  website: string;
  yield: number;
  ytdReturn: number;
  zip: string;

  constructor() { }

}

export interface IHoldingInfo {
  '52WeekChange': number;
  SandP52WeekChange: number;
  address1: string;
  address2: string;
  algorithm: string;
  annualHoldingsTurnover: string;
  annualReportExpenseRatio: string;
  ask: number;
  askSize: number;
  averageDailyVolume10Day: number;
  averageVolume: number;
  averageVolume10days: number;
  beta: number;
  beta3Year: number;
  bid: number;
  bidSize: number;
  bookValue: number;
  category: string;
  circulatingSupply: number;
  city: string;
  companyOfficers: Array<string>;
  country: string;
  currency: string;
  dateShortInterest: number;
  dayHigh: number;
  dayLow: number;
  dividendRate: number;
  dividendYield: number;
  earningsQuarterlyGrowth: number;
  enterpriseToEbitda: number;
  enterpriseToRevenue: number;
  enterpriseValue: number;
  exDividendDate: number;
  exchange: string;
  exchangeTimezoneName: string;
  exchangeTimezoneShortName: string;
  expireDate: string;
  fiftyDayAverage: number;
  fiftyTwoWeekHigh: number;
  fiftyTwoWeekLow: number;
  fiveYearAverageReturn: number;
  fiveYearAvgDividendYield: number;
  floatShares: number;
  forwardEps: number;
  forwardPE: number;
  fromCurrency: string;
  fullTimeEmployees: number;
  fundFamily: string;
  fundInceptionDate: string;
  gmtOffSetMilliseconds: number;
  heldPercentInsiders: number;
  heldPercentInstitutions: number;
  industry: string;
  isEsgPopulated: boolean;
  lastCapGain: number;
  lastDividendValue: number;
  lastFiscalYearEnd: number;
  lastMarket: string;
  lastSplitDate: number;
  lastSplitFactor: string;
  legalType: string;
  'logo_url': string;
  longBusinessSummary: string;
  longName: string;
  market: string;
  marketCap: number;
  maxAge: number;
  maxSupply: number;
  messageBoardId: string;
  morningStarOverallRating: number;
  morningStarRiskRating: number;
  mostRecentQuarter: number;
  navPrice: number;
  netIncomeToCommon: number;
  nextFiscalYearEnd: number;
  open: number;
  openInterest: number;
  payoutRatio: number;
  pegRatio: number;
  phone: string;
  previousClose: number;
  priceHint: number;
  priceToBook: number;
  priceToSalesTrailing12Months: number;
  profitMargins: number;
  quoteType: string;
  regularMarketDayHigh: number;
  regularMarketDayLow: number;
  regularMarketOpen: number;
  regularMarketPreviousClose: number;
  regularMarketPrice: number;
  regularMarketVolume: number;
  revenueQuarterlyGrowth: number;
  sector: string;
  sharesOutstanding: number;
  sharesPercentSharesOut: number;
  sharesShort: number;
  sharesShortPreviousMonthDate: number;
  sharesShortPriorMonth: number;
  shortName: string;
  shortPercentOfFloat: number;
  shortRatio: number;
  startDate: string;
  state: string;
  strikePrice: number;
  symbol: string;
  threeYearAverageReturn: number;
  toCurrency: string;
  totalAssets: number;
  tradeable: boolean;
  trailingAnnualDividendRate: number;
  trailingAnnualDividendYield: number;
  trailingEps: number;
  trailingPE: number;
  twoHundredDayAverage: number;
  volume: number;
  volume24Hr: number;
  volumeAllCurrencies: number;
  website: string;
  yield: number;
  ytdReturn: number;
  zip: string;
}